admin-server style refactor #2733
