+++ src/scene_server/admin_server/authsynchronizer/README.md
