+++ src/scene_server/admin_server/authsynchronizer/handler/classification.go
@@ -19,43 +19,44 @@ import (
 	"configcenter/src/auth/extensions"
 	authmeta "configcenter/src/auth/meta"
 	"configcenter/src/common/blog"
-	"configcenter/src/scene_server/admin_server/synchronizer/meta"
-	"configcenter/src/scene_server/admin_server/synchronizer/utils"
+	"configcenter/src/scene_server/admin_server/authsynchronizer/meta"
+	"configcenter/src/scene_server/admin_server/authsynchronizer/utils"
 )
 
-// HandleSetSync do sync set of one business
-func (ih *IAMHandler) HandleSetSync(task *meta.WorkRequest) error {
+// HandleModuleSync do sync module of one business
+func (ih *IAMHandler) HandleClassificationSync(task *meta.WorkRequest) error {
 	businessSimplify := task.Data.(extensions.BusinessSimplify)
 	header := utils.NewAPIHeaderByBusiness(&businessSimplify)
 
 	// step1 get instances by business from core service
-	businessID := businessSimplify.BKAppIDField
-	sets, err := ih.authManager.CollectSetByBusinessID(context.Background(), *header, businessID)
+	bizID := businessSimplify.BKAppIDField
+	classifications, err := ih.authManager.CollectClassificationByBusinessIDs(context.Background(), *header, bizID)
 	if err != nil {
-		blog.Errorf("get set by business id:%d failed, err: %+v", businessID, err)
+		blog.Errorf("collect classifications by business id failed, err: %+v", err)
 		return err
 	}
-	if len(sets) == 0 {
-		blog.Infof("no set found for business: %d", businessID)
+	if len(classifications) == 0 {
+		blog.Infof("no classifications found for business: %d", bizID)
 		return nil
 	}
-	resources := ih.authManager.MakeResourcesBySet(*header, authmeta.EmptyAction, businessID, sets...)
-	if len(resources) == 0 && len(sets) > 0 {
-		blog.Errorf("make iam resource for sets %+v return empty", sets)
+	resources := ih.authManager.MakeResourcesByClassifications(*header, authmeta.EmptyAction, bizID, classifications...)
+	if len(resources) == 0 && len(classifications) > 0 {
+		blog.Errorf("make iam resource for classifications %+v return empty", classifications)
 		return nil
 	}
 
-	// step2 get set by business from iam
+	// step2 get classifications by business from iam
 	rs := &authmeta.ResourceAttribute{
 		Basic: authmeta.Basic{
-			Type: authmeta.ModelSet,
+			Type: authmeta.ModelClassification,
 		},
-		BusinessID: businessSimplify.BKAppIDField,
-		Layers:     make([]authmeta.Item, 0),
+		SupplierAccount: "",
+		BusinessID:      businessSimplify.BKAppIDField,
+		Layers:          make([]authmeta.Item, 0),
 	}
 
-	taskName := fmt.Sprintf("sync set for business: %d", businessSimplify.BKAppIDField)
-	iamIDPrefix := "set"
+	taskName := fmt.Sprintf("sync classifications for business: %d", businessSimplify.BKAppIDField)
+	iamIDPrefix := ""
 	skipDeregister := false
 	return ih.diffAndSync(taskName, rs, iamIDPrefix, resources, skipDeregister)
 }
