+++ src/scene_server/admin_server/authsynchronizer/handler/module.go
@@ -19,44 +19,44 @@ import (
 	"configcenter/src/auth/extensions"
 	authmeta "configcenter/src/auth/meta"
 	"configcenter/src/common/blog"
-	"configcenter/src/scene_server/admin_server/synchronizer/meta"
-	"configcenter/src/scene_server/admin_server/synchronizer/utils"
+	"configcenter/src/scene_server/admin_server/authsynchronizer/meta"
+	"configcenter/src/scene_server/admin_server/authsynchronizer/utils"
 )
 
 // HandleModuleSync do sync module of one business
-func (ih *IAMHandler) HandleClassificationSync(task *meta.WorkRequest) error {
+func (ih *IAMHandler) HandleModuleSync(task *meta.WorkRequest) error {
 	businessSimplify := task.Data.(extensions.BusinessSimplify)
 	header := utils.NewAPIHeaderByBusiness(&businessSimplify)
 
 	// step1 get instances by business from core service
 	bizID := businessSimplify.BKAppIDField
-	classifications, err := ih.authManager.CollectClassificationByBusinessIDs(context.Background(), *header, bizID)
+	modules, err := ih.authManager.CollectModuleByBusinessIDs(context.Background(), *header, bizID)
 	if err != nil {
-		blog.Errorf("collect classifications by business id failed, err: %+v", err)
+		blog.Errorf("collect module by business id failed, err: %+v", err)
 		return err
 	}
-	if len(classifications) == 0 {
-		blog.Infof("no classifications found for business: %d", bizID)
+	if len(modules) == 0 {
+		blog.Infof("no modules found for business: %d", bizID)
 		return nil
 	}
-	resources := ih.authManager.MakeResourcesByClassifications(*header, authmeta.EmptyAction, bizID, classifications...)
-	if len(resources) == 0 && len(classifications) > 0 {
-		blog.Errorf("make iam resource for classifications %+v return empty", classifications)
+	resources := ih.authManager.MakeResourcesByModule(*header, authmeta.EmptyAction, bizID, modules...)
+	if len(resources) == 0 && len(modules) > 0 {
+		blog.Errorf("make iam resource for modules %+v return empty", modules)
 		return nil
 	}
 
-	// step2 get classifications by business from iam
+	// step2 get modules by business from iam
 	rs := &authmeta.ResourceAttribute{
 		Basic: authmeta.Basic{
-			Type: authmeta.ModelClassification,
+			Type: authmeta.ModelModule,
 		},
 		SupplierAccount: "",
 		BusinessID:      businessSimplify.BKAppIDField,
 		Layers:          make([]authmeta.Item, 0),
 	}
 
-	taskName := fmt.Sprintf("sync classifications for business: %d", businessSimplify.BKAppIDField)
-	iamIDPrefix := ""
+	taskName := fmt.Sprintf("sync module for business: %d", businessSimplify.BKAppIDField)
+	iamIDPrefix := "module"
 	skipDeregister := false
 	return ih.diffAndSync(taskName, rs, iamIDPrefix, resources, skipDeregister)
 }
