+++ src/scene_server/admin_server/authsynchronizer/meta/types.go
