+++ src/scene_server/admin_server/authsynchronizer/utils/types.go
