+++ src/scene_server/admin_server/migrate.go
@@ -19,7 +19,7 @@ import (
 	"runtime"
 
 	"github.com/spf13/pflag"
-	
+
 	"configcenter/src/common"
 	"configcenter/src/common/blog"
 	"configcenter/src/common/types"
@@ -41,13 +41,13 @@ func main() {
 	op.AddFlags(pflag.CommandLine)
 
 	if err := command.Parse(os.Args); err != nil {
-		fmt.Fprintf(os.Stderr, "%v\n", err)
+		fmt.Fprintf(os.Stderr, "parse arguments failed, %v\n", err)
 		os.Exit(1)
 	}
 	util.InitFlags()
 
 	if err := app.Run(context.Background(), op); err != nil {
-		fmt.Fprintf(os.Stderr, "%v\n", err)
+		fmt.Fprintf(os.Stderr, "run app failed, %v\n", err)
 		blog.Errorf("process stopped by %v", err)
 		blog.CloseLogs()
 		os.Exit(1)
